Password update docs
