+++ components/doc/password/floatlabeldoc.js
@@ -1,6 +1,8 @@
 import { DocSectionCode } from '@/components/doc/common/docsectioncode';
 import { DocSectionText } from '@/components/doc/common/docsectiontext';
+import { FloatLabel } from '@/components/lib/floatlabel/FloatLabel';
 import { Password } from '@/components/lib/password/Password';
+import Link from 'next/link';
 import { useState } from 'react';
 
 export function FloatLabelDoc(props) {
@@ -8,24 +10,25 @@ export function FloatLabelDoc(props) {
 
     const code = {
         basic: `
-<span className="p-float-label">
+<FloatLabel>
     <Password inputId="password" value={value} onChange={(e) => setValue(e.target.value)} />
     <label htmlFor="password">Password</label>
-</span>
+</FloatLabel>
         `,
         javascript: `
 import React, { useState } from "react";
 import { Password } from 'primereact/password';
+import { FloatLabel } from 'primereact/floatlabel';
 
 export default function FloatLabelDemo() {
     const [value, setValue] = useState('');
 
     return (
         <div className="card flex justify-content-center">
-            <span className="p-float-label">
+            <FloatLabel>
                 <Password inputId="password" value={value} onChange={(e) => setValue(e.target.value)} />
                 <label htmlFor="password">Password</label>
-            </span>
+            </FloatLabel>
         </div>
     )
 }
@@ -33,16 +36,17 @@ export default function FloatLabelDemo() {
         typescript: `
 import React, { useState } from "react";
 import { Password } from 'primereact/password';
+import { FloatLabel } from 'primereact/floatlabel';
 
 export default function FloatLabelDemo() {
     const [value, setValue] = useState<string>('');
 
     return (
         <div className="card flex justify-content-center">
-            <span className="p-float-label">
+            <FloatLabel>
                 <Password inputId="password" value={value} onChange={(e: React.ChangeEvent<HTMLInputElement>) => setValue(e.target.value)} />
                 <label htmlFor="password">Password</label>
-            </span>
+            </FloatLabel>
         </div>
     )
 }
@@ -52,13 +56,15 @@ export default function FloatLabelDemo() {
     return (
         <>
             <DocSectionText {...props}>
-                <p>A floating label appears on top of the input field when focused.</p>
+                <p>
+                    A floating label appears on top of the input field when focused. Visit <Link href="/floatlabel">FloatLabel</Link> documentation for more information.
+                </p>
             </DocSectionText>
             <div className="card flex justify-content-center">
-                <span className="p-float-label">
+                <FloatLabel>
                     <Password inputId="password" value={value} onChange={(e) => setValue(e.target.value)} />
                     <label htmlFor="password">Password</label>
-                </span>
+                </FloatLabel>
             </div>
             <DocSectionCode code={code} />
         </>
