--- components/doc/password/form/hookformdoc.js
@@ -1,188 +1,0 @@
-import { DocSectionCode } from '@/components/doc/common/docsectioncode';
-import { DocSectionText } from '@/components/doc/common/docsectiontext';
-import { Button } from '@/components/lib/button/Button';
-import { Password } from '@/components/lib/password/Password';
-import { Toast } from '@/components/lib/toast/Toast';
-import { classNames } from '@/components/lib/utils/Utils';
-import { useRef } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-
-export function HookFormDoc(props) {
-    const toast = useRef(null);
-
-    const show = () => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: form.getValues('value') });
-    };
-
-    const defaultValues = { value: '' };
-    const form = useForm({ defaultValues });
-    const errors = form.formState.errors;
-
-    const onSubmit = (data) => {
-        data.value && show();
-
-        form.reset();
-    };
-
-    const getFormErrorMessage = (name) => {
-        return errors[name] ? <small className="p-error">{errors[name].message}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    const code = {
-        basic: `
-<Toast ref={toast} />
-<Controller
-    name="value"
-    control={form.control}
-    rules={{ required: 'Password is required.' }}
-    render={({ field, fieldState }) => (
-        <>
-            <label htmlFor={field.name} className={classNames({ 'p-error': errors.value })}>
-                Password
-            </label>
-            <Password id={field.name} {...field} inputRef={field.ref} className={classNames({ 'p-invalid': fieldState.error })} feedback={false} />
-            {getFormErrorMessage(field.name)}
-        </>
-    )}
-/>
-<Button label="Submit" type="submit" icon="pi pi-check" />
-        `,
-        javascript: `
-import React, { useRef } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-import { Button } from 'primereact/button';
-import { Toast } from 'primereact/toast';
-import { classNames } from 'primereact/utils';
-import { Password } from "primereact/password";
-
-export default function HookFormDoc() {
-    const toast = useRef(null);
-
-    const show = () => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: form.getValues('value') });
-    };
-
-    const defaultValues = { value: '' };
-    const form = useForm({ defaultValues });
-    const errors = form.formState.errors;
-
-    const onSubmit = (data) => {
-        data.value && show();
-
-        form.reset();
-    };
-
-    const getFormErrorMessage = (name) => {
-        return errors[name] ? <small className="p-error">{errors[name].message}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    return (
-        <div className="card flex justify-content-center">
-            <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-column gap-2">
-                <Toast ref={toast} />
-                <Controller
-                    name="value"
-                    control={form.control}
-                    rules={{ required: 'Password is required.' }}
-                    render={({ field, fieldState }) => (
-                        <>
-                            <label htmlFor={field.name} className={classNames({ 'p-error': errors.value })}>
-                                Password
-                            </label>
-                            <Password id={field.name} {...field} inputRef={field.ref} className={classNames({ 'p-invalid': fieldState.error })} feedback={false} />
-                            {getFormErrorMessage(field.name)}
-                        </>
-                    )}
-                />
-                <Button label="Submit" type="submit" icon="pi pi-check" />
-            </form>
-        </div>
-    )
-}
-        `,
-        typescript: `
-import React, { useRef } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-import { Button } from 'primereact/button';
-import { Toast } from 'primereact/toast';
-import { classNames } from 'primereact/utils';
-import { Password } from "primereact/password";
-
-export default function HookFormDoc() {
-    const toast = useRef(null);
-
-    const show = () => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: form.getValues('value') });
-    };
-
-    const defaultValues = { value: '' };
-    const form = useForm({ defaultValues });
-    const errors = form.formState.errors;
-
-    const onSubmit = (data) => {
-        data.value && show();
-
-        form.reset();
-    };
-
-    const getFormErrorMessage = (name) => {
-        return errors[name] ? <small className="p-error">{errors[name].message}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    return (
-        <div className="card flex justify-content-center">
-            <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-column gap-2">
-                <Toast ref={toast} />
-                <Controller
-                    name="value"
-                    control={form.control}
-                    rules={{ required: 'Password is required.' }}
-                    render={({ field, fieldState }) => (
-                        <>
-                            <label htmlFor={field.name} className={classNames({ 'p-error': errors.value })}>
-                                Password
-                            </label>
-                            <Password id={field.name} {...field} inputRef={field.ref} className={classNames({ 'p-invalid': fieldState.error })} feedback={false} />
-                            {getFormErrorMessage(field.name)}
-                        </>
-                    )}
-                />
-                <Button label="Submit" type="submit" icon="pi pi-check" />
-            </form>
-        </div>
-    )
-}
-        `
-    };
-
-    return (
-        <>
-            <DocSectionText {...props}>
-                <p>
-                    <a href="https://react-hook-form.com/">React Hook Form</a> is another popular React library to handle forms.
-                </p>
-            </DocSectionText>
-            <div className="card flex justify-content-center">
-                <form onSubmit={form.handleSubmit(onSubmit)} className="flex flex-column gap-2">
-                    <Toast ref={toast} />
-                    <Controller
-                        name="value"
-                        control={form.control}
-                        rules={{ required: 'Password is required.' }}
-                        render={({ field, fieldState }) => (
-                            <>
-                                <label htmlFor={field.name} className={classNames({ 'p-error': errors.value })}>
-                                    Password
-                                </label>
-                                <Password id={field.name} {...field} inputRef={field.ref} className={classNames({ 'p-invalid': fieldState.error })} feedback={false} />
-                                {getFormErrorMessage(field.name)}
-                            </>
-                        )}
-                    />
-                    <Button label="Submit" type="submit" icon="pi pi-check" />
-                </form>
-            </div>
-            <DocSectionCode code={code} dependencies={{ 'react-hook-form': '^7.39.4' }} />
-        </>
-    );
-}
